requirements.txt
numpy~=2.1.3
pillow~=11.0.0